+++ crates/tests/src/lib.rs
@@ -4,6 +4,7 @@
 #![expect(clippy::disallowed_methods, clippy::needless_raw_string_hashes)]
 
 mod check;
+mod docs;
 mod hover;
 mod misc;
 mod repo;