+++ crates/jsonnet-statics/src/check.rs
@@ -276,7 +276,6 @@ fn get_subscript(
   idx: Expr,
 ) -> ty::Ty {
   let idx_expr = idx.unwrap_or(expr);
-  // TODO handle unions
   match st.tys.data(on_ty).clone() {
     ty::Data::Prim(ty::Prim::Any) => ty::Ty::ANY,
     ty::Data::Prim(_) | ty::Data::Fn(_) => {
@@ -320,9 +319,10 @@ fn get_subscript(
         }
       }
     }
-    ty::Data::Union(_) => {
-      st.unify(on.unwrap_or(expr), ty::Ty::ARRAY_OR_OBJECT, on_ty);
-      ty::Ty::ANY
+    ty::Data::Union(tys) => {
+      let iter = tys.into_iter().map(|on_ty| get_subscript(st, ar, on_ty, idx_ty, expr, on, idx));
+      let res = ty::Data::Union(iter.collect());
+      st.tys.get(res)
     }
   }
 }